tests: remove mock_seccomp rust test.
The new seccompiler library will no longer make the seccomp
primitives available to Firecracker. This commit removes the use
of seccomp library in the rust integration tests.

This test was verifying the behaviour of a 'too restrictive' seccomp
filter in Firecracker. This is also tested in test_custom_seccomp.py
integration test.

Signed-off-by: alindima <alindima@amazon.com>